Authentication provider now checks if JT token really exist, if not or you put a fake like 'banaan' it will be logged out
