+++ src/provider/AuthenticationProvider/AuthenticationProvider.jsx
@@ -1,33 +1,36 @@
 import React, { createContext, useState, useEffect, useContext } from 'react';
-import {jwtDecode} from "jwt-decode";
-import axios from "axios";
-import HomePage from "../../pages/HomePage/HomePage.jsx";
+import {jwtDecode} from 'jwt-decode';
+import axios from 'axios';
+
 const AuthenticationContext = createContext();
 export const useAuthentication = () => {
     return useContext(AuthenticationContext);
 };
 
-
-//AuthenticationProvider controleert of een gebruiker is ingelogd en of de JWT-token aanwezig is op elke pagina
 export const AuthenticationProvider = ({ children }) => {
     const [isLoggedIn, setIsLoggedIn] = useState(false);
     const [username, setUsername] = useState('');
     const [userAuthorities, setUserAuthorities] = useState();
 
-
     useEffect(() => {
-        // Nodig om te kijken of JWT token bestaat in local storage
         const jwtToken = localStorage.getItem('jwtToken');
         if (jwtToken) {
-            setIsLoggedIn(true);
-            //Decodeer token om user / gebruikersnaam informatie te decrypten
-            const decodedToken = jwtDecode(jwtToken);
+            try {
+                // Decodeer de token en controleer vervaldatum
+                const decodedToken = jwtDecode(jwtToken);
 
-            // Extract gebruikersnaam van token;
-            const { sub: username } = decodedToken;
-
-            // Zet de gebruikersnaam state
-            setUsername(username);
+                // Controleer of de token is verlopen
+                if (decodedToken.exp * 1000 < Date.now()) {
+                    handleLogout();
+                } else {
+                    // Token is nog geldig, zet de state
+                    setIsLoggedIn(true);
+                    setUsername(decodedToken.sub);
+                }
+            } catch (error) {
+                console.error('Error decoding token', error);
+                handleLogout();
+            }
         }
     }, []);
 
@@ -37,7 +40,6 @@ export const AuthenticationProvider = ({ children }) => {
                 if (username) {
                     const response = await axios.get(`http://localhost:8080/users/${username}/authorities`);
                     const authorities = response.data.map(authority => {
-                        //Gebruikersrol definieren
                         if (authority.authority === 'ROLE_USER') {
                             return 'Gebruiker';
                         } else if (authority.authority === 'ROLE_ADMIN') {
@@ -57,12 +59,10 @@ export const AuthenticationProvider = ({ children }) => {
     }, [username]);
 
     const handleLogout = () => {
-        // Perform logout actions here (e.g., clear authentication state)
         setIsLoggedIn(false);
-        localStorage.removeItem('jwtToken'); // Clear token from localStorage
-
-        // Navigate to the home page
-        window.location.href = '/'; // Replace '/' with the actual URL of your home page
+        setUsername('');
+        localStorage.removeItem('jwtToken');
+        window.location.href = '/'; // Verwijs naar de homepagina
     };
 
     return (
